Updated Code With Header Files
Used struct in the .h files to declare all variables within the pump and waterlevel sensor code. The .cpp has the code to run the function and base.ino declares input/output, loops of function, and creates instances of the structs. Verified this code on ArduinoIDE and tested on SimulIDE and both are good.