Anthonyw298 (#9)
* Added a fan cycle code where fan A is on 1s theb B is on 1s then both are one 1s then both are off 3s before cycling thru again

Also reoutlined the code by using struct to make it more nice to view

* Added Code for Pump.ino Unit testing utilizing the same structure code as fans.ino unit testing code

THis code will initialize the pump to move with the testing of the pump speed and being able to adjust it

* Revised Base.ino Code

Here i revised the code where I labeled various struct types that initializes the related variables&values within the function that corresponds to the section of the water pump system. I then installed the various functions to run the smart system after these structures. FInally I call these functions and run it for a loop. This code includes the fan cycle and water pump cycle using PWM.

* Unit Testing for pump.ino

Added Code for Pump.ino Unit testing utilizing the same structure code as fans.ino unit testing code. This code will initialize the pump to move with the testing of the pump speed and being able to adjust it

* Variable Grouping Of Pump

Took away the fan and other outside parts of code and just focused on varuable grouping the pump variables into a structure type and then kept the original code and functions for the rest

* Variable Grouping Of Pump

Took away the fan and other outside parts of code and just focused on varuable grouping the pump variables into a structure type and then kept the original code and functions for the rest

* Test

* Test

* Base.ino Variable Grouping

Here is the actual revised code. based off the previous comments you had, I added a structure that grouped all the variables for the pump and kept the rest of the code the same.

* Updated Code With Header Files

Used struct in the .h files to declare all variables within the pump and waterlevel sensor code. The .cpp has the code to run the function and base.ino declares input/output, loops of function, and creates instances of the structs. Verified this code on ArduinoIDE and tested on SimulIDE and both are good.